+++ Assets/Scripts/MessagingSystem/Messages/DamageDisplayEvent.cs
@@ -4,11 +4,11 @@ namespace MyMessagingSystem
 {
     public struct DamageDisplayEvent
     {
-        public float damageAmount;
+        public int damageAmount;
         public Transform displayLocation;
         public Color damageColor;
 
-        public DamageDisplayEvent(float damageAmount, Transform displayLocation, Color damageColor)
+        public DamageDisplayEvent(int damageAmount, Transform displayLocation, Color damageColor)
         {
             this.damageAmount = damageAmount;
             this.displayLocation = displayLocation;
